parent component for categoryPractice and categoryQuiz
