+++ mute-mate.client/src/categoryPractise.css
@@ -22,10 +22,12 @@ body {
 .option-img {
   max-width: 16rem;
   max-height: 16rem;
-  margin-bottom: 1rem;
-  border: 0, 3px solid black;
+  /* margin-bottom: 1rem; */
+  border: 0 3px solid black;
   border-radius: 10px;
   padding-left: 2rem;
+  background-color: #FEFEE3;
+  margin: 1rem;
 }
 .page-wrapper {
   display: flex;
@@ -43,4 +45,6 @@ body {
 }
 .choose-text {
   text-align: center;
+  color:  #FEFEE3;
+  padding-bottom: 1rem;
 }
