+++ mute-mate.client/src/pages/categoryQuiz.jsx
@@ -1,5 +1,14 @@
+import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
+import CategoryPractiseQuiz from "../components/CategoryPractiseQuiz";
 function categoryQuiz() {
-  return <></>;
+  return (
+    <>
+      <CategoryPractiseQuiz 
+      message="Let´s quiz"
+      linkTo={"/quiz"}/>
+    </>
+  );
 }
 
 export default categoryQuiz;
